src/modules/motion_capture/requirements.txt
# 动捕数据采集模块依赖项
protobuf>=3.20.0
pyzmq>=22.0.0
pyside6>=6.0.0
